requirements.txt
lalsuite
astropy
gwpy
python-ligo-lw
numpy
pandas
